pytest: runtime/platform refactor
We had a lot of tests in platform/ that were supposed to be in runtime/

The reason it's better they are in runtime is because they get tested
against k8s earlier -- which means we detect problems earlier.
platform/ tests really are supposed to check APIs not runtime details.
