+++ python/tests/runtime/lateness/test_issue4895.py
@@ -6,7 +6,7 @@ from feldera.testutils import (
     FELDERA_TEST_NUM_HOSTS,
 )
 from tests import TEST_CLIENT
-from tests.platform.helper import PipelineTestCase
+from tests.helper import PipelineTestCase
 from feldera.enums import PipelineStatus
 from feldera.runtime_config import RuntimeConfig
