+++ python/tests/runtime/test_nowstream.py
@@ -8,7 +8,7 @@ from feldera.testutils import (
     FELDERA_TEST_NUM_HOSTS,
 )
 from tests import TEST_CLIENT
-from tests.platform.helper import PipelineTestCase
+from tests.helper import PipelineTestCase
 from feldera.enums import PipelineStatus
