+++ python/tests/runtime/test_output_buffer_checkpoint.py
@@ -26,7 +26,7 @@ from feldera.testutils import (
 )
 from tests import TEST_CLIENT
 
-from .helper import gen_pipeline_name
+from tests.helper import gen_pipeline_name
 
 
 @enterprise_only
